+++ src/components/Form.vue
@@ -95,7 +95,7 @@
 </template>
 
 <script>
-import Button from "../components/Button.vue";
+import Button from "./Button.vue";
 
 export default {
   components: { Button },
